feat: save search in cookies
