+++ starWars/package.json
@@ -4,21 +4,21 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "test": "jest --coverage --collectCoverageFrom='src/**/*.{ts,tsx}'",
+    "test": "jest --coverage --collectCoverageFrom='./**/*.tsx'",
     "start": "next start",
     "dev": "next dev",
     "build": "next build",
     "lint": "next lint",
     "lint:fix": "next lint --fix",
-    "format:fix": "prettier --write ./src",
-    "format": "prettier --check ./src",
+    "format:fix": "prettier --write",
+    "format": "prettier --check",
     "preview": "vite preview",
-    "prepare": "cd .. && husky starWars/.husky"
+    "prepare": "cd .. && husky starWars/.husky",
+    "test:watch": "jest --watchAll"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^2.2.6",
     "@testing-library/user-event": "^14.5.2",
-    "jest-environment-jsdom": "^29.7.0",
     "mvs": "^1.0.0",
     "next": "^14.2.5",
     "react": "^18.3.1",
@@ -31,7 +31,7 @@
   },
   "devDependencies": {
     "@next/eslint-plugin-next": "^14.2.5",
-    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/jest-dom": "^6.4.8",
     "@testing-library/react": "^16.0.0",
     "@types/jest": "^29.5.12",
     "@types/react": "^18.3.3",
@@ -51,6 +51,7 @@
     "husky": "^9.0.11",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "jest-fetch-mock": "^3.0.3",
     "lint-staged": "^15.2.7",
     "msw": "^2.3.4",
