--- starWars/src/test/__mocks__/hooks.ts
@@ -1,5 +1,0 @@
-export const useGetDetailsQuery = jest.fn();
-export const useGetFilmsQuery = jest.fn();
-export const useGetSearchAndPageQuery = jest.fn();
-export const useAppDispatch = jest.fn();
-export const useAppSelector = jest.fn();
